+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Rename module: `Containers` -> `VectorsOfVectors`.
+
 ## [0.32.12] - 2025-12-21
 
 ### Added